fix: separate wish request DTOs
