+++ src/main/java/gift/api/wishlist/dto/WishDeleteRequest.java
@@ -1,0 +1,10 @@
+package gift.api.wishlist.dto;
+
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Positive;
+
+public record WishDeleteRequest(
+    @NotNull(message = "Product id is mandatory")
+    @Positive(message = "Product id must be greater than zero")
+    Long productId
+) {}
